+++ core/migrations/release_4_1_0.php
@@ -16,7 +16,7 @@ class release_4_1_0 extends migration
 {
 	public static function depends_on(): array
 	{
-		return ['\phpbbgallery\core\migrations\gallery_album_bbcode'];
+		return ['\phpbbgallery\core\migrations\image_card_bbcode_id'];
 	}
 
 	public function update_data(): array